third_party/blink/web_tests/external/wpt/css/css-contain/content-visibility/content-visibility-090-expected.txt

This is a testharness.js-based test.
[FAIL] Content Visibility: scrollLeft/scrollTop/scrollWidth/scrollHeight measure correctly
  assert_equals: left expected 25 but got 0
Harness: the test ran to completion.
